added delivery sort

+++ app/src/main/java/com/wallpad/notice/repository/Repository.java
@@ -130,7 +130,7 @@ public class Repository {
         return deliveries;
     }
 
-    public void readNoticeDelivery(int id) {
+    public void readNoticeDelivery(long id) {
         executorService.execute(() -> deliveryDao.updateRead(id, true));
     }
